+++ NEWS.md
@@ -18,6 +18,7 @@
 * `any_table()`: If the custom NA symbol is set to a number, Excels number stored as text error is now ignored. (12.02.2026)
 * `transpose_plus()`: Took out a debug print I forgot in the function. (12.02.2026)
 * `summarise_plus()`: Fixed formats not matching data when computing percentages, if numeric values are stored as character. (18.02.2026)
+* `if.()`, `else_if.()`, `else.()`: When passing a vector to a new variable, the functions don't error any more if there are NA values. (22.02.2026)
 
 ### Additionally